Improve input validation and error handling across various components
- Added checks for finite values in `bm_algo_clock_drift_feed`, `bm_algo_signal_quality`, and other functions to prevent NaN/Inf propagation.
- Enhanced validation in configuration functions such as `bm_estimation_fusion_validate_config` and `bm_motion_profile_validate_config` to reject invalid modes and parameters.
- Introduced safety measures in control loops and state machines to reset integrators on faults or disabled states.
- Implemented additional telemetry updates for fault conditions in sensor quality and navigation components.
- Improved handling of edge cases in scheduling and logging mechanisms to ensure thread safety and prevent resource leaks.
- Added comments for clarity and future improvements in components lacking state machines or fault handling.

+++ Source/core/bm_shell.c
@@ -240,7 +240,7 @@ static void _modal_exit(bm_shell_t *shell) {
  */
 static void _tab_complete(bm_shell_t *shell) {
     const char *match = NULL;
-    uint8_t     n_match = 0;
+    uint16_t    n_match = 0;
     uint8_t     help_registered = 0;
 
     /* 只补全行首命令词：已进入参数区（前缀含空白）时不补全 */
